+++ src/components/Loading/Loading.scss
@@ -1,10 +1,20 @@
+.loading {
+  position: fixed;
+  width: 100%;
+  height: 100%;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  z-index: 2;
+}
+
 .spinner {
   display: flex;
   align-items: center;
   justify-content: center;
   height: 100%;
   width: 100%;
-  background: rgba(0, 0, 0, 0.7);
   position: fixed;
 
   img {
