+++ src/components/Loading/Loading.tsx
@@ -1,0 +1,11 @@
+import loading from "../../assets/loading.svg";
+
+const Loading = () => (
+  <div className="loading">
+    <div className="spinner">
+      <img src={loading} alt="Loading" />
+    </div>
+  </div>
+);
+
+export default Loading;
